Update to move to the new cn-test branch.
