+++ lib/worker.js
@@ -239,7 +239,7 @@ function updateShareStats() {
                 Object.keys(identifiers).forEach(function (key) {
                     cache_updates[key + '_identifiers'] = identifiers[key];
                 });
-                cache_updates['minerList'] = globalMinerList;
+                cache_updates.minerList = globalMinerList;
                 global.database.bulkSetCache(cache_updates);
                 callback(null);
             });
